Get, delete, post, patch working to users table
change sqlalchemy to flask.ext.sqlalchemy
implemented callbacks to api users
