+++ templates/login.html
@@ -1,5 +1,5 @@
-<div id="login_form">
-    <form name="f1" method="post" action="login.php" id="f1">
+<div>
+    <form name="f1" method="post" action="login">
         <table>
             <tr>
                 <td class="f1_label">User Name :</td><td><input type="text" name="username" value="" />
