+++ templates/post_users.html
@@ -1,0 +1,31 @@
+<div>
+    <form method="post" action="/postUser"  enctype='application/json'>
+        <table>
+            <tr>
+                <td class="f1_label">User Name :</td><td><input type="text" name="username" value="" />
+                </td>
+            </tr>
+            <tr>
+                <td class="f1_label">Password  :</td><td><input type="password" name="password" value=""  />
+                </td>
+            </tr>
+            <tr>
+            <tr>
+                <td class="f1_label">clid  :</td><td><input type="clid" name="clid" value=""  />
+                </td>
+            </tr>
+            <tr>
+                <td class="f1_label">balance  :</td><td><input type="balance" name="balance" value=""  />
+                </td>
+            </tr>
+            <tr>
+                <td class="f1_label">admin  :</td><td><input type="admin" name="admin" value=""  />
+                </td>
+            </tr>
+                <td>
+                    <input type="submit" name="addUser" value="create" style="font-size:18px; " />
+                </td>
+            </tr>
+        </table>
+    </form>
+</div>
